Split giant cookie.ts into multiple files. (#296)
* Quiet IDE warnings about TS v5 deprecations.

* Add eslint warning for files over 500 lines.

* Split giant cookie.ts into multiple files.

* Update main export.
