+++ lib/__tests__/canonicalDomain.spec.ts
@@ -1,4 +1,4 @@
-import { canonicalDomain } from '../cookie'
+import { canonicalDomain } from '../cookie/canonicalDomain'
 
 // port of tests/domain_and_path_test.js (canonicalDomain tests for domain normalization)
 describe('canonicalDomain', () => {
