+++ lib/__tests__/cookie.spec.ts
@@ -29,7 +29,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { Cookie } from '../cookie'
+import { Cookie } from '../cookie/cookie'
 
 jest.useFakeTimers()
