+++ lib/__tests__/pathMatch.spec.ts
@@ -1,4 +1,4 @@
-import { pathMatch } from '../cookie'
+import { pathMatch } from '../pathMatch'
 
 // port of tests/domain_and_path_test.js (path match tests)
 describe('pathMatch', () => {
