+++ package.json
@@ -76,8 +76,8 @@
   "bugs": {
     "url": "https://github.com/salesforce/tough-cookie/issues"
   },
-  "main": "./dist/cookie.js",
-  "types": "./dist/cookie.d.ts",
+  "main": "./dist/cookie/index.js",
+  "types": "./dist/cookie/index.d.ts",
   "files": [
     "dist/*.js",
     "dist/*.d.ts"
